# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Public content API for the VietPress EU news portal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description: string | null;
}

export interface Country {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  code: string | null;
}

export interface Author {
  id: number;
  name: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  avatar: string | null;
}

export interface Article {
  id: number;
  title: string;
  slug: string;
  summary: string;
  content: string;
  /** @nullable */
  coverImage: string | null;
  category: Category;
  country: Country | null;
  author: Author | null;
  /** @nullable */
  sourceName: string | null;
  /** @nullable */
  sourceUrl: string | null;
  /** @nullable */
  editor: string | null;
  /** @nullable */
  publishedAt: string | null;
  status: string;
  featured: boolean;
  breakingNews: boolean;
  views: number;
}

export interface Event {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  startDate: string;
  /** @nullable */
  endDate: string | null;
  /** @nullable */
  location: string | null;
  /** @nullable */
  image: string | null;
  /** @nullable */
  registrationUrl: string | null;
  eventType: string;
}

export type HomepagePayloadEuCountries = {[key: string]: Article[]};

export interface HomepagePayload {
  breakingNews: Article[];
  featured: Article | null;
  mostRead: Article[];
  selected: Article[];
  euCountries: HomepagePayloadEuCountries;
  vietnam: Article[];
  world: Article[];
  business: Article[];
  features: Article[];
  golfEvents: Event[];
  communityEvents: Event[];
}

export interface ArticleListResponse {
  items: Article[];
  page: number;
  pageSize: number;
  total: number;
}

export interface ViewResponse {
  views: number;
}

export interface NewsletterInput {
  email: string;
}

export interface NewsletterSubscription {
  id: number;
  email: string;
  active: boolean;
  subscribedAt: string;
}

export type PageParameter = number;

export type PageSizeParameter = number;

export type CategoryParameter = string;

export type CountryParameter = string;

export type ListArticlesParams = {
/**
 * @minimum 1
 */
page?: PageParameter;
/**
 * @minimum 1
 * @maximum 50
 */
pageSize?: PageSizeParameter;
category?: CategoryParameter;
country?: CountryParameter;
};

export type SearchArticlesParams = {
/**
 * @minLength 1
 */
q: string;
/**
 * @minimum 1
 */
page?: PageParameter;
/**
 * @minimum 1
 * @maximum 50
 */
pageSize?: PageSizeParameter;
};

export type ListEventsParams = {
eventType?: string;
};