# lib/api-zod/src/generated/types/articleListResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Public content API for the VietPress EU news portal
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from './article';

export interface ArticleListResponse {
  items: Article[];
  page: number;
  pageSize: number;
  total: number;
}