# lib/api-zod/src/generated/types/author.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Public content API for the VietPress EU news portal
 * OpenAPI spec version: 0.1.0
 */

export interface Author {
  id: number;
  name: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  avatar: string | null;
}