# lib/api-zod/src/generated/types/newsletterInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Public content API for the VietPress EU news portal
 * OpenAPI spec version: 0.1.0
 */

export interface NewsletterInput {
  email: string;
}